rename src/cmd -> src/commands
