+++ src/commands/check.rs
