+++ src/lib.rs
@@ -1,3 +1,3 @@
-pub mod cmd;
+pub mod commands;
 pub mod services;
 pub mod rust_analyzer;
